Revert "added step definitions for getStationDetails and setOperatinStatus"
