+++ src/test/java/org/example/SD_RegisterationDetails.java
@@ -1,0 +1,48 @@
+package org.example;
+
+import io.cucumber.java.en.And;
+import io.cucumber.java.en.Given;
+import io.cucumber.java.en.Then;
+import io.cucumber.java.en.When;
+
+public class SD_RegisterationDetails {
+    @Given("I am a new user")
+    public void iAmANewUser() {
+    }
+
+    @When("I enter my name, email, and password into the registration form")
+    public void iEnterMyNameEmailAndPasswordIntoTheRegistrationForm() {
+    }
+
+    @Then("my registration should be processed")
+    public void myRegistrationShouldBeProcessed() {
+    }
+
+    @And("I should receive a unique customer ID")
+    public void iShouldReceiveAUniqueCustomerID() {
+    }
+
+    @Given("I have successfully registered")
+    public void iHaveSuccessfullyRegistered() {
+    }
+
+    @When("the registration is completed")
+    public void theRegistrationIsCompleted() {
+    }
+
+    @Then("I should receive a confirmation email with my unique customer ID")
+    public void iShouldReceiveAConfirmationEmailWithMyUniqueCustomerID() {
+    }
+
+    @Given("I have not entered all required fields")
+    public void iHaveNotEnteredAllRequiredFields() {
+    }
+
+    @When("I attempt to register")
+    public void iAttemptToRegister() {
+    }
+
+    @Then("the system should display an error message asking for the missing details")
+    public void theSystemShouldDisplayAnErrorMessageAskingForTheMissingDetails() {
+    }
+}
